connect: move other structs into model.rs
